move test fns in to test_lca_functions
